fix per comments
